Revert "Cursor Pos now a Double Var"
This reverts commit 5e8cdcbf179e6f2329ba9e32f391d3ae17cb5ea4.

+++ Leviathan/Input.cpp
@@ -37,19 +37,19 @@ bool Input::isKeyPressed(int keycode) {
 
 }
 
-double Input::getCursorX() {
+float Input::getCursorX() {
 
 	return cursorPosX;
 
 }
 
-double Input::getCursorY() {
+float Input::getCursorY() {
 
 	return cursorPosY;
 
 }
 
-double* Input::getCursorArray() {
+float* Input::getCursorArray() {
 
 	float returnArr[2]; //Array to return values in
